Add files via upload
This is the file that contains main or driver code of the program. It creates an array of object add than call verification and main_menu of different users. It than use the concept of polymorphism (function overriding via virtual functions and calls the respective pointed objects of other classes).